+++ component/backend/Model/Subscriptions.php
@@ -303,11 +303,6 @@ class Subscriptions extends DataModel
 
 				return;
 			}
-
-			// Otherwise we have to do a relation filter against the user relation, filtering by business name or VAT number
-			$this->whereHas('user', function (\JDatabaseQuery $q) use($search) {
-				$q->where($q->qn('businessname') . ' LIKE ' . $q->q("%$search%"));
-			});
 		}
 	}